dbt-acquisition/requirements.txt
pydantic==2.13.3
python-dotenv==1.2.2
PyYAML==6.0.3
Requests==2.33.1
SQLAlchemy==2.0.49
sqlglot==28.10.1
